backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
Pygments==2.13.0
pytz==2022.6
sqlparse==0.4.3
